requirements.txt
python-telegram-bot[webhooks]>=21.5,<22
python-dotenv>=1.0.0
tzdata>=2024.1
google-api-python-client>=2.130.0
google-auth>=2.29.0
